= Moved `gmGenIndex` into it's own file
