+++ src/ECS/mod.rs
@@ -12,6 +12,7 @@ mod world;
 mod resource;
 mod builders;
 mod dispatcher;
+mod misc;
 
 use comp::*;
 use storage::*;
@@ -21,29 +22,7 @@ use world::*;
 use resource::*;
 use builders::*;
 use dispatcher::*;
-
-#[derive(Clone, Copy)]
-pub struct gmGenIndex<T: Sized>{
-    pub id: gmID,
-    pub gen: gmGen,
-    pub entry: Option<T>
-}
-impl<T: Sized> gmGenIndex<T>{
-    pub fn new(IN_id: gmID, IN_entry: Option<T>) -> Self{
-        Self{
-            id: IN_id,
-            gen: 0,
-            entry: IN_entry,
-        }
-    }
-    pub fn set(&mut self, IN_entry: T){
-        self.entry = Some(IN_entry);
-    }
-    pub fn unset(&mut self){
-        self.entry = None;
-        self.gen += 1;
-    }
-}
+use misc::*;
 
 mod tests{
     use event::*;
